a little modify

+++ src/main/java/com/nju/edu/community/entity/Post.java
@@ -36,9 +36,9 @@ public class Post {
     private String contentUrl;//文章内容URL,存储在OSS上
 
     private long publishTime;//发布时间
-    private long visits;//浏览数
-    private long remarkNum;//评论数
-    private long interestNum;//被收藏的次数
+    private int visits;//浏览数
+    private int remarkNum;//评论数
+    private int interestNum;//被收藏的次数
 
     @OneToMany(cascade = CascadeType.ALL)
     private List<Remark> remarkList;//文章下的评论列表
@@ -46,19 +46,8 @@ public class Post {
     public Post(String author){
         SimpleDateFormat df = new SimpleDateFormat("yyyyMMddHHmmss");//设置日期格式
         String currentTime=df.format(new Date());// new Date()为获取当前系统时间
-//        long currentTime=new Date().getTime();
         this.pid =currentTime+author.substring(0,author.indexOf('@'));
         this.author=author;
         this.state=PostState.Draft;
     }
-
-//    public Post(String pid, String author, String title, PostTag postTag, String brief_intro, String contentUrl){
-//        this.pid = pid;
-//        this.author=author;
-//        this.title = title;
-//        this.postTag = postTag;
-//        this.brief_intro=brief_intro;
-//        this.contentUrl = contentUrl;
-//    }
-
 }
